MAJOR update, switching back from OO style
gfortran did not seem to be ready for full Object Oriented programming style.
Frequently it would have a problem compiling or running something that appeared correct
and worked fine in ifort.  So this is refactoring to use a more procedural/imperitive style.

Time module remains OO, and will stay that way for now.

This commit also includes a basic functionality to read in a netcdf file
(or files/variables) using the gcm module.