Speed up quantile mapping by a few orders of magnitude
Changed the search algorithm for the quantile to map to from an O(n) to a binary search O(log(n)).
Also improved the test_qm and added a verbose option to test_qm.

Also added INC_NCAR/LIB_NCAR to the compile/link lines, may want to rethink how that is applied later.