improved minval handling for normalization
There was an issue related to normalizing the predictor data based on the training data

Most steps in that normalization step use pre-calculated values, so it wasn’t a problem
but there is a last step where the minimum value in the dataset is subtracted
(to get rid of negative values in case one wanted to take a sqrt)
That step uses the minimum value of the normalization data.

Now a minval field is added to the variable datatype, allocated and initialized
in the IO code, and updated in the normalization code after removing mean and stddev
but before removing the minval itself.