Updates to quantile mapping code to smooth it a bit
This *should* particularly be helpful for extreme values.  Previously
the QM code was just picking off n values from the CDFs of input and output
now it computes a mean value over a window from the CDFs. Importantly
for the most extreme values, instead of using the absolute high and low
values from each distribution, it computes a mean over a very small window.